reactor_jobm::placement

Trait PlacementManager

pub trait PlacementManager {
    // Required method
    fn place(&self, op_info: &LogicalOp) -> impl Iterator<Item = PhysicalOp>;
}

Takes logical Op and places it on single or multiple Nodes. Returns list of Physical operator where a logical operator is placed

Required Methods

fn place(&self, op_info: &LogicalOp) -> impl Iterator<Item = PhysicalOp>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.